+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Server has `getConfig` method to property access to the server configuration.
 - Option to change location of source styles that are merged into the application styles.
 - Option to disable automatically socket connection.
-- Preparation for plugins.
+- Plugins can be registered to the server.
 - Definition of accepted locales.
 - Option to set locale from the application.
 - Server can register middlewares.